Enforce blockReader interface  (#1023)
